+++ theme_xtream/doc/RELEASE_NOTES.md
@@ -1,7 +1,7 @@
-## Module <xtream_theme>
+## Module <theme_xtream>
 
 #### 11.11.2024
 #### Version 17.0.1.0.0
 #### ADD
 
-- Initial commit for Lytics-Theme Xtream Fashion
+- Initial commit for Theme Xtream Fashion